git-svn-id: http://miranda.googlecode.com/svn/branches/stable@14047 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/Gadu-Gadu/image.c
@@ -910,7 +910,7 @@ BOOL gg_img_hasextension(const char *filename)
 		char *imgtype = strrchr(filename, '.');
 		if (imgtype != NULL)
 		{
-			int len = strlen(imgtype);
+			size_t len = strlen(imgtype);
 			imgtype++;
 			if (len == 4 && (_stricmp(imgtype, "bmp") == 0 ||
 							 _stricmp(imgtype, "gif") == 0 ||
